Added provisional workaround for registering GReQL functions if such a function is called from within an Eclipse Plug-In:
The problem is that with the plug-in class loader, getResource() in registerAllFunctions does only return an URL beginning with "de/uni-koblenz/...", which cannot be processed by registerFunctionsInJar() and registerFunctionsInDirectory(). Within the latter, an exception was caused because the "entries" variable is null.

The workaround is implemented by method registeredFixedFunctionSet(), which registers the functions specified in the "functions" field.

TODO: Improve this solution once someone knows how to get the full path for the jgralab jar from within a plug-in ;) .