Fix logging.
The problem was that a Handler has a Level, too, and that's completely
orthogonal to the Logger's level, because many loggers may use the same
handler.  Additionally, the default level of Handlers is INFO.  Now the jgralab
root loggers ConsoleHandler uses ALL, and all child loggers use their parent's
level.