Fix major inconsistency in type-specific edge/incidence iteration code generation.
With changeset e3163363ddec, Daniel improved the edge iteration stuff in this
way:

  - Iterate globally (eSeq):

    myGraph.getFirstEdge().getNextEdge().getNextEdge()...

  - Iterate incidences of a vertex:

    myVertex.getFirstIncidence().getNextIncidence().getNextIncidence()...

However, Daniel missed the next-methods stuff in the code generator.  So if you
had an edge class Link, then the global iteration using type-specific methods
was

  myGraph.getFirstLink().getNextLinkInGraph().getNextLinkInGraph()...

and the incidence iteration

  myVertex.getFirstLinkIncidence().getNextLink().getNextLink()...

which is totally inconsistent.

This changeset adapts the code generator to stick to the naming convention of
the generic iteration code and the getFirst-methods.  So now you have the
global iteration code

  myGraph.getFirstLink().getNextLink().getNextLink()...

and the incidence iteration

  myVertex.getFirstLinkIncidence().getNextLinkIncidence().getNextLinkIncidence()...

WARNING: This changes the interface quite a bit if type specific methods were
used.  So probably, some stuff has to be adapted.