added assert statements to methods in Vertex and Edge for asserting that the current graph element is in TC before actually executing the methods getFirst/LastIncidence, getNext/PrevVertex/Edge/Incidence
Restored the behavior of AStar in case of passing start and target vertices that are not part of the subgraph (now TC): an IllegalArgumentException is thrown.

Restored the behavior of Search algorithms when called with root vertex not in subgraph (now TC): the vertex is just ignored and nothing happens (the algorithm doesn't even start).