Add hashCode() & equals() to OrderedPSet.
That fixes a bunch of optimizer tests which failed cause the sets were not
equal, because the missing equality implementation of OrderedPSet.

Also, changed a bunch of optimizer test queries to returns sets instead of
lists.  There are some optimizations (like shuffeling variables) that don't
preserve the order of the results.  Daniel and I agree that this is just fine.

Fixed another test that failed due to a missing function.