Deprecate Vertex#isValidAlpha/Omega(Edge).
Get rid of their implementation in the generated code, but implement them in
VertexBaseImpl via (now fast) schema lookup.

What's left to be done: Basically, these methods should be removed.  Currently,
they are used like that:

An edge is created, and when associating it with their alpha/omega vertex,
these methods tests for schema conformity.  That's pretty convoluted.  This
check should be done *before* an edge is being created, not afterwards.  That
would simplify a lot of code.  Furthemore, the GENERIC impl has no way to
create an edge without alpha/omega.

There are some troubles with the factories, which completely rely on the
interfaces and have no idea about the vertex/edge classes.  Where can the check
via the schema be stuffed in so that it works ideally for all implementation
variants?