Allow for limiting comprehensions.
Added an optional maxCount Expression (EdgeClass IsMaxCountExpressionOf) to
Comprehension in the GReQL schema, which must result in an integral number and
specifies the maximum number of elements to be calculated and reported.

In the concrete syntax, this is reflected by the new keywords reportSetN,
reportListN, and reportMapN, which get the limit as first expression followed
by a COMMA, followed by the usual result definition.

Example: the first 10 primes in the range [500..1000]

  from c: list(500..1000)
  with not(exists d: list(2..c-1) @ c % d = 0)
  reportListN 10: c end

This feature is important for GReTL in-place transformations, where you need to
find some arbitrary occurence of elements in the graph matching some pattern,
which are then replaced.  Because the replacement might impact later matches,
calculation of all matches (with a usual FWR) is a waste.  You cannot be
completely sure that later matches found initially are still valid.