Added missing testcases.
Completed testcases for graph markers as functions.
Added all missing testcases to the corresponding test suites.
Detected and removed a bug in the function-related iterators of BitSet graphmarkers. Now all vertices in the graph belong to the function domain (matching the behavior of isDefined), not only the marked graph elements.