Implemented Unset Attribute Feature
Implemented a new feature that allows checking if an attribute is unset, which
is needed by FunnyQT bidirectional transformations in order for the
transformation to determine what attributes of existing elements it is allowed
to modify.  The only user-visible new method is:

  boolean AttributedElement.isUnsetAttribute(String name)

This method returns true it the given attribute hasn't been set in terms of
setAttribute() already.  Default values declared in the schema also count as
unset, unless they've been explicitly set in terms of setAttribute().

There are no effects wrt. getAttribute().  That still returns false for unset
booleans, 0 for unset integers, null for unset strings, lists, maps, or the
default value declared for the schema element.

When serializing to TG, unset attribute values are written as `u` similar to
`n` for null values, and `t` and `f` for booleans.  When loading a TG file,
attributes with value `u` stay unset and getAttribute() will return whatever is
the standard value (0, 0.0, false, null) or the schema declared default value.

Of course that feature has also some little drawbacks:

  - every attributed element now also has a BitSet tracking the attributes that
    have already been set.

  - loading a generic graph might be a bit slower, cause now we cannot skip
    initializing default values when the graph is loading.

While I've been at it, I also deleted some transaction impl leftovers, fixed
and refactored some tests, simplified the code generator here and there, and of
course, there are new tests for that new unset attributes feature.